backend/fast_api_predict/requirements.txt
fastapi==0.62.0
python-multipart
uvicorn[standard]
Pillow